Extract configuration methods
