+++ 02-UIKit/Today/Today/DetailViewController/ReminderViewController+CellConfiguration.swift
@@ -1,0 +1,34 @@
+//
+//  ReminderViewController+CellConfiguration.swift
+//  Today
+//
+//  Created by Imam on 08/10/23.
+//
+
+import UIKit
+
+extension ReminderViewController {
+    func defaultConfiguration(for cell: UICollectionViewListCell, at row: Row) -> UIListContentConfiguration {
+        var contentConfiguration = cell.defaultContentConfiguration()
+        contentConfiguration.text = text(for: row)
+        contentConfiguration.textProperties.font = UIFont.preferredFont(forTextStyle: row.textStyle)
+        contentConfiguration.image = row.image
+        return contentConfiguration
+    }
+    
+    func headerConfiguration(for cell: UICollectionViewListCell, with title: String) -> UIListContentConfiguration {
+        var contentConfiguration = cell.defaultContentConfiguration()
+        contentConfiguration.text = title
+        return contentConfiguration
+    }
+    
+    func text(for row: Row) -> String? {
+        switch row {
+        case .date: return reminder.dueDate.dayText
+        case .notes: return reminder.notes
+        case .time: return reminder.dueDate.formatted(date: .omitted, time: .shortened)
+        case .title: return reminder.title
+        default: return nil
+        }
+    }
+}
